removed binaries and added gzip extraction headers for a simpler process
this should allow the app to be installed and not have issues when installed without using trollstore. This version should be compatible with altstore and other 3rd party app signers. opnly the app binary now needs to be re-signed. Like it should be. Also removed binpack.tar and changed it to a gz file, compressing the tar file from 90Mb to 30Mb 1/3 the size.